refactor :: email패키지와 auth패키지 분리
