+++ src/main/java/com/example/climserver/domain/auth/api/AuthController.java
@@ -5,9 +5,6 @@ import com.example.climserver.domain.auth.dto.request.LoginRequest;
 import com.example.climserver.domain.auth.dto.request.RefreshTokenRequest;
 import com.example.climserver.domain.auth.dto.request.SignupRequest;
 import com.example.climserver.domain.auth.dto.response.TokenResponse;
-import com.example.climserver.domain.auth.email.application.EmailService;
-import com.example.climserver.domain.auth.email.application.VerificationCodeService;
-import jakarta.mail.MessagingException;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
