+++ src/main/java/com/example/climserver/domain/email/application/EmailService.java
@@ -1,4 +1,4 @@
-package com.example.climserver.domain.auth.email.application;
+package com.example.climserver.domain.email.application;
 
 import jakarta.mail.MessagingException;
 import jakarta.mail.internet.MimeMessage;
