+++ src/main/java/com/example/climserver/global/security/jwt/JwtProperties.java
@@ -2,6 +2,7 @@ package com.example.climserver.global.security.jwt;
 
 import lombok.Getter;
 import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
 
 import java.util.Base64;
